Implemented the GCC task that parses module interface units. Refactored a little bit the code, decoupling the logic of generating this commands from the type that acts as the serde parser for the cpp compiler
