+++ zork++/src/lib/compiler/commands.rs
@@ -9,6 +9,11 @@ use color_eyre::{eyre::Context, Result};
 /// Executes a new [`std::process::Command`] configured according the choosen
 /// compiler and the current operating system
 pub fn execute_command(compiler: &CppCompiler, arguments: Vec<String>) -> Result<()> {
+    log::info!(
+        "[{compiler}] - Executing command {:?}",
+        format!("{} {}", compiler.get_driver(), arguments.join(" "))
+    );
+
     let process = if compiler.eq(&CppCompiler::MSVC) {
         Command::new( // TODO The initialization process + cache process MUST dynamically get this path and store it in cache
             "C:\\Program Files\\Microsoft Visual Studio\\2022\\Community\\VC\\Auxiliary\\Build\\vcvars64.bat"
@@ -26,11 +31,6 @@ pub fn execute_command(compiler: &CppCompiler, arguments: Vec<String>) -> Result
             .with_context(|| format!("[{compiler}] - Command {:?} failed!", arguments.join(" ")))?
     };
 
-    log::info!(
-        "[{compiler}] - Command {:?}\nresult: {:?}",
-        arguments.join(" "),
-        process
-    );
-
+    log::info!("[{compiler}] - Result: {:?}", process);
     Ok(())
 }
